ci: Make publish helm chart and create release workflows sequential (#814)
**Reason for Change**:
As part of tighten the security for Kaito pipelines, we enforced
approval-based for all workflows that perform writing to the repo. One
of the side effects of this was a required approval for the auto
triggered 'pages build and deployment' workflow which was missing - and
it will get canceled when creating a new release - and caused the helm
github page repo to not get updated (#808).

As a solution to that, we will change the Make publish helm chart and
create release workflows to run sequentially instead of parallel, so new
release won't get created until all (including github pages) are updated
successfully.


**Requirements**

- [ ] added unit tests and e2e tests (if applicable).

**Issue Fixed**:
<!-- If this PR fixes GitHub issue 4321, add "Fixes #4321" to the next
line. -->

**Notes for Reviewers**:

Signed-off-by: Heba Elayoty <heelayot@microsoft.com>