feat: parse supported_models.yaml in plugin pkg before starting manager (#1044)
**Reason for Change**:
<!-- What does this PR improve or fix in Kaito? Why is it needed? -->

Treats `presets/workspace/models/supported_models.yaml` as the source of
truth for handling model-specific static information such as model
versions, container image tags, and whether download is necessary or
not.

**Requirements**

- [x] added unit tests and e2e tests (if applicable).

**Issue Fixed**:
<!-- If this PR fixes GitHub issue 4321, add "Fixes #4321" to the next
line. -->

N/A

**Notes for Reviewers**:

TODO in a follow-up PR: Eliminate any tags map in every `model.go` and
treat the tag in `supported_models.yaml` as the source of truth.

---------

Signed-off-by: Ernest Wong <chwong719@gmail.com>