fix: ensureNodePlugins does not work (#1070)
**Reason for Change**:
<!-- What does this PR improve or fix in Kaito? Why is it needed? -->
`resources.CheckNvidiaPlugin(ctx, nodeObj)` checked `nodeObj`, which is
a cached value. Retry is no use.

Another problem is that `return nil` in `default` condition will always
be executed no matter found or not found. I think this is not what we
want. This fix assume that we only `return nil` when found.

**Requirements**

- [ ] added unit tests and e2e tests (if applicable).

**Issue Fixed**:
<!-- If this PR fixes GitHub issue 4321, add "Fixes #4321" to the next
line. -->

**Notes for Reviewers**:

---------

Co-authored-by: Fei Guo <guofei@microsoft.com>