feat: add update/delete document api's and handling for faiss (#1090)
**Reason for Change**:
This adds update/delete document api's and handling for FAISS. This
required a wrapper around the LlamaIndex FaissVectorStore and wrapping
our initially used FlatL2 faiss index with a IDMap index to support
delete operations. Base functionality is still the same with the wrapped
index, we just now have support for document mapping to id's in the
IDMap index.

**Requirements**

- [x] added unit tests and e2e tests (if applicable).

**Issue Fixed**:
<!-- If this PR fixes GitHub issue 4321, add "Fixes #4321" to the next
line. -->

**Notes for Reviewers**:

---------

Co-authored-by: Fei Guo <guofei@microsoft.com>