docs: add keda scaler for kaito workloads proposal (#1237)
**Reason for Change**:
<!-- What does this PR improve or fix in KAITO? Why is it needed? -->
According to the Jorge's comments, I proposed an auto-scaler solution
based on keda system.

Compared to `LLMAutoScaler` proposal in #1172 , I think both of them
have the same capabilities, but the user inferface is different, so we
need to decide which one to use from a production perspective. and we
will have a meeting to discuss these two proposals.

**Requirements**

- [ ] added unit tests and e2e tests (if applicable).

**Issue Fixed**:
<!-- If this PR fixes GitHub issue 4321, add "Fixes #4321" to the next
line. -->

**Notes for Reviewers**:

---------

Signed-off-by: rambohe-ch <rambohe.ch@gmail.com>