chore: linter/formatter for python code (#1351)
**Reason for Change**:
Lints and formats the existing python code for improved maintenance and
readability.

**Requirements**

- [ ] added unit tests and e2e tests (if applicable).

**Issue Fixed**:
Fixes #1343 

**Notes for Reviewers**:
> presets/workspace/inference/vllm/inference_api.py:167:42: F821
Undefined name `executor`

This error seems to be a false positive, but I can't confirm it.

You can run the linter through `ruff check` and the formatter through
`ruff format`.
It would benefit the project to move all python dependency management to
[uv](https://docs.astral.sh/uv/).

---------

Signed-off-by: Farooq Ameen <102469559+farooqameen@users.noreply.github.com>
Co-authored-by: jerryzhuang <zhuangqhc@gmail.com>