fix: ResourceReady condition is never set to true for BYO [release-0.7]  (#1549)
**Reason for Change**:
<!-- What does this PR improve or fix in KAITO? Why is it needed? -->

The resource ready condition is never set to true after the refactor as
it compares the # ready node claims to the # nodes needed rather than
the # node claims needed. This breaks as if 1 node is needed and one
existing node is brought, it will expect 1 node claim that will never be
provisioned as the # nodes is already met.

**Requirements**

- [ ] added unit tests and e2e tests (if applicable).

**Issue Fixed**:
<!-- If this PR fixes GitHub issue 4321, add "Fixes #4321" to the next
line. -->
Fixes #1544 

**Notes for Reviewers**:

**Reason for Change**:
<!-- What does this PR improve or fix in KAITO? Why is it needed? -->

**Requirements**

- [ ] added unit tests and e2e tests (if applicable).

**Issue Fixed**:
<!-- If this PR fixes GitHub issue 4321, add "Fixes #4321" to the next
line. -->

**Notes for Reviewers**: